feat(context): movendo funções para context
Compartilhando funções de incremento atraves da context
